blooms-classifier/requirements.txt
torch
transformers
datasets
accelerate>=0.26.0
sentence-transformers
scikit-learn
pandas
numpy>=2.1.3,<2.2
scipy
matplotlib
tqdm
textstat
# spacy is optional; omitted for Python 3.13 wheels. Install manually if desired.
# spacy
# liwc-text-analysis
evaluate
pyyaml
jupyter
weightwatcher
rapidfuzz
beautifulsoup4
trafilatura
requests
duckduckgo-search
jsonlines
joblib
plotly
nbformat
nbconvert
papermill
protobuf
